Refactor: Merge redundant DiagnosticRequestID components and fix input crash
- Merged `DiagnosticRequestIDMapper` into `DiagnosticRequestIDManager` to serve as the single source of truth for CAN ID mapping logic.
- Unified SharedPreferences key prefixes and consolidated configuration mapping methods (`getMapping()`, `updateSettings()`, `reset()`).
- Standardized the mapping loop limit (`DRI_MAX_MAPPINGS_ALLOWED`) to 100 to prevent potential data loss.
- Fixed an `IllegalArgumentException` (WordIterator invalid offset) during mapping creation by removing the conflicting `android:textAllCaps="true"` attribute from the `TextInputEditText` in `dialog_dri_edit.xml`. Capitalization remains safely handled in the bottom sheet logic.
