+++ app/src/main/java/org/obd/graphs/preferences/dri/DiagnosticRequestIdFragment.kt
@@ -24,6 +24,8 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
+import org.obd.graphs.DiagnosticMappingItem
+import org.obd.graphs.DiagnosticRequestIDManager
 import org.obd.graphs.R
 
 class DiagnosticRequestIdFragment : Fragment() {
