+++ app/src/main/res/layout/dialog_dri_edit.xml
@@ -40,8 +40,7 @@
         <com.google.android.material.textfield.TextInputEditText
             android:id="@+id/etValue"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:textAllCaps="true" />
+            android:layout_height="wrap_content" />
     </com.google.android.material.textfield.TextInputLayout>
 
     <com.google.android.material.button.MaterialButton
